Successfully pass user as prop to EventPage from app
