+++ client/components/users/UserProfile.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Link } from 'react-router';
 import ListOfEvents from './ListOfEvents.js';
-import { IndexLink } from 'react-router';
+import { Link } from 'react-router';
 
 require('./../../styles/styles.css');
 
@@ -9,32 +9,6 @@ export default class UserProfile extends React.Component {
 
   constructor(props) {
     super(props);
-
-    this.state = {
-      username: '',
-      events: []
-    }
-  }
-
-  componentDidMount() {
-    this.fetchProfile();
-  }
-
-  fetchProfile() {
-    $.ajax({
-      type: 'GET',
-      url: '/api/users',
-      dataType: 'json',
-      success: function(data) {
-        this.setState({
-          username: data.local.username,
-          events: data.events
-        });
-      }.bind(this),
-      fail: function(err) {
-        console.error(err);
-      }
-    });
   }
 
   render() {
